- FBX loading corrections using submeshes
- JPEG fix
- logging improvement

+++ DirectZobEngine/MaterialManager.h
@@ -12,7 +12,7 @@ public:
 
 	const Material* LoadMaterial(const std::string& name, const Vector3* ambientColor, const Vector3* diffuseColor, const char* textureFile = NULL);
 	const Material* GetMaterial(const std::string& name) const;
-	const int GetNbTextures() const { return m_materials.size(); }
+	const int GetNbTextures() const { return (int)m_materials.size(); }
 	const Material* GetMaterial(const int i) const;
 	void LoadMaterials(std::string& path, std::string& file);
 	void UnloadAll();
