+++ DirectZobEngine/ZobObject.h
@@ -34,7 +34,7 @@ public:
 	inline const std::string& GetName() const { return m_name; }
 	inline void SetName(const std::string &name) { m_name = name; }
 	const void GetFullNodeName(std::string &fullName) const;
-	inline const int GetNbChildren() const { return m_children.size(); }
+	inline const int GetNbChildren() const { return (int)m_children.size(); }
 	void RemoveChildReference(const ZobObject* z);
 	void AddChildReference(ZobObject* z);
 	int GetChildPosition(const ZobObject* z);
